@@ -1,7 +1,7 @@
-from typing import List, Type
+from typing import Type
 from abc import ABC
 
 
 class BaseExecutorClassLoader(ABC):
-    def load(self) -> List[Type]:
+    def load(self) -> list[Type]:
         raise NotImplementedError()
