@@ -1,9 +1,9 @@
-from typing import List, Type
+from typing import Type
 
 from ...task_templates.class_loader import BaseTaskTemplateClassLoader
 from ...task_templates.system_command import SystemCommandTaskTemplate
 
 
 class ClassLoader(BaseTaskTemplateClassLoader):
-    def load(self) -> List[Type]:
+    def load(self) -> list[Type]:
         return [SystemCommandTaskTemplate]
