@@ -1,10 +1,9 @@
-from typing import Dict, List
-
 from ...task_templates.task_template import BaseTaskTemplate
+from ...typedefs import StringKvDict
 
 
 class SystemCommandTaskTemplate(BaseTaskTemplate):
     """A task template for system command"""
     kind: str = "system_command"
-    cmd: List[str] = []
-    env: Dict[str, str] = {}
+    cmd: list[str] = []
+    env: StringKvDict = {}
