@@ -1,4 +1,4 @@
-from typing import List, Type
+from typing import Type
 from abc import ABC
 
 
@@ -6,5 +6,5 @@ class BaseTaskTemplateClassLoader(ABC):
     """
     Task template class loaders return a list of task template classes to main module.
     """
-    def load(self) -> List[Type]:
+    def load(self) -> list[Type]:
         raise NotImplementedError()
