@@ -1,12 +1,11 @@
-from typing import Dict, List
-
 from .task_template import BaseTaskTemplate
+from ..typedefs import StringKvDict
 
 
 class SystemCommandTaskTemplate(BaseTaskTemplate):
     """A task template for system command"""
     kind: str = "system_command"
-    cmd: List[str] = []
+    cmd: list[str] = []
     """A command to execute"""
-    env: Dict[str, str] = {}
+    env: StringKvDict = {}
     """Environment variables"""
